+++ packages/api/CHANGELOG.md
@@ -1,5 +1,50 @@
 # Changelog
 
+## [0.9.0](https://github.com/F3-Nation/f3-nation/compare/pkg-api@0.8.0...pkg-api@0.9.0) (2026-09-10)
+
+
+### Features
+
+* **api,admin,db,shared:** add OAuth client admin UI ([#957](https://github.com/F3-Nation/f3-nation/issues/957)) ([f09c9c5](https://github.com/F3-Nation/f3-nation/commit/f09c9c5a5d0921833eec876fbcee102999ceed37))
+
+
+### Bug Fixes
+
+* **api:** replace fixed-depth ancestor-active check with recursive CTE ([#967](https://github.com/F3-Nation/f3-nation/issues/967)) ([cd6e1fe](https://github.com/F3-Nation/f3-nation/commit/cd6e1fe29dd270b19cbe7f463fb58cfac5f3aa58))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @acme/auth bumped to 0.2.4
+    * @acme/db bumped to 0.6.0
+    * @acme/shared bumped to 0.3.0
+    * @acme/validators bumped to 0.4.1
+
+## [0.8.0](https://github.com/F3-Nation/f3-nation/compare/pkg-api@0.7.0...pkg-api@0.8.0) (2026-09-08)
+
+
+### Features
+
+* **admin:** add event tags and instances management modals ([#493](https://github.com/F3-Nation/f3-nation/issues/493)) ([f80ce91](https://github.com/F3-Nation/f3-nation/commit/f80ce91edc00de91d2f1c8c12a557d883fadfac8))
+
+
+### Bug Fixes
+
+* **api:** add .output() schemas to event-tag router ([#902](https://github.com/F3-Nation/f3-nation/issues/902)) ([6aa08ca](https://github.com/F3-Nation/f3-nation/commit/6aa08caf611d8731483916561a27a728975b1ec8))
+* **api:** stop crupdate resetting isActive/highlight/isPrivate on partial updates ([#900](https://github.com/F3-Nation/f3-nation/issues/900)) ([f89b7f7](https://github.com/F3-Nation/f3-nation/commit/f89b7f7e8c0ccb3341ee98c644d7d98836180a4a))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @acme/auth bumped to 0.2.3
+    * @acme/db bumped to 0.5.0
+    * @acme/shared bumped to 0.2.0
+    * @acme/validators bumped to 0.4.0
+
 ## [0.7.0](https://github.com/F3-Nation/f3-nation/compare/pkg-api@0.6.0...pkg-api@0.7.0) (2026-08-18)